+++ services/wiki/src/__tests__/resources/byTopic.test.ts
@@ -1,7 +1,7 @@
 import { Category, Topic, User, Prisma } from '@prisma/client'
 import supertest from 'supertest'
 import { expect, test, describe, beforeAll, afterAll } from 'vitest'
-import { server, testUserData } from '../globalSetup'
+import { server, testCategoryData, testUserData } from '../globalSetup'
 import { pathRoot } from '../../routes/routes'
 import { prisma } from '../../prisma/client'
 import { resourceGetSchema } from '../../schemas'
@@ -14,7 +14,7 @@ let userWithNoName: User | null
 
 beforeAll(async () => {
   const testCategory = (await prisma.category.findUnique({
-    where: { slug: 'testing' },
+    where: { slug: testCategoryData.slug },
   })) as Category
 
   testTopic = (await prisma.topic.findUnique({
@@ -94,35 +94,6 @@ describe('GET /resources/topic/:topicId', () => {
   test('should fail if topic ID does not exist in database ', async () => {
     const response = await supertest(server).get(
       `${baseUrl}/this-topicId-does-not-exist`
-    )
-
-    expect(response.status).toBe(404)
-    expect(response.body.message).toEqual('Topic not found')
-  })
-})
-
-describe('GET /v1/resources/topic/slug/:slug', () => {
-  const baseUrl = `${pathRoot.v1.resources}/topic/slug`
-
-  test('should respond with OK status if topic slug exists in database and return an array of resources associated with the topic slug  ', async () => {
-    const response = await supertest(server).get(`${baseUrl}/${testTopic.slug}`)
-
-    expect(response.status).toBe(200)
-    expect(response.body.resources).toBeInstanceOf(Array)
-    expect(response.body.resources.length).toBeGreaterThan(0)
-    response.body.resources.forEach((resource: any) => {
-      expect(() => resourceGetSchema.parse(resource)).not.toThrow()
-      expect(
-        resource.topics.some(
-          (topic: any) => topic.topic.slug === testTopic.slug
-        )
-      ).toBe(true)
-    })
-  })
-
-  test('should fail if topic slug does not exist in database ', async () => {
-    const response = await supertest(server).get(
-      `${baseUrl}/this-topicSlug-does-not-exist`
     )
 
     expect(response.status).toBe(404)