+++ services/wiki/src/openapi/routes/resources/list.ts
@@ -1,7 +1,7 @@
 import { registry } from '../../registry'
 import { z } from '../../zod'
 import { pathRoot } from '../../../routes/routes'
-import { resourceGetSchema, resourcesGetParamsSchema } from '../../../schemas'
+import { resourceGetSchema, resourcesListParamsSchema } from '../../../schemas'
 import {
   invalidTokenResponse,
   zodValidationErrorResponse,
@@ -15,7 +15,7 @@ registry.registerPath({
     'Returns a collection of resources. Filters by resource type, topic name and category slug are optional. Resources that match all filters are fetched.',
   summary: 'Returns a collection of resources',
   request: {
-    query: resourcesGetParamsSchema,
+    query: resourcesListParamsSchema,
   },
   responses: {
     200: {